Fix #1875: Inject ptvsd into subprocesses Fix #1901: Race between IDE and server connecting to the adapter
Use "evaluate" request without "frameId" to import ptvsd into the debuggee as soon as server attaches.

If the adapter is spawned by the server, wait for server to connect if IDE connects first.

Fix test_subprocess to use the new test API, and add a check for ptvsd being loaded in the child process.